refactor: board 패키지 분리
