+++ src/main/kotlin/minesweeper/domain/generator/PositionGenerator.kt
@@ -1,6 +1,6 @@
 package minesweeper.domain.generator
 
-import minesweeper.domain.BoardMeta
+import minesweeper.domain.board.BoardMeta
 import minesweeper.domain.field.Position
 
 fun interface PositionGenerator {
