+++ src/test/kotlin/minesweeper/domain/board/BoardMetaTest.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.board
 
 import io.kotest.assertions.throwables.shouldThrow
 import org.junit.jupiter.api.Test
