+++ src/test/kotlin/minesweeper/domain/board/BoardSizeTest.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.board
 
 import io.kotest.assertions.throwables.shouldThrow
 import org.junit.jupiter.params.ParameterizedTest
