+++ src/test/kotlin/minesweeper/domain/board/BoardTest.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.board
 
 import io.kotest.assertions.assertSoftly
 import io.kotest.matchers.shouldBe
